Add files via upload
Exploratory Data Analysis - Retail

Retail is a dynamic and competitive sector that generates vast amounts of data every day. This internship task will provide you with the opportunity to dive deep into this data and extract meaningful information to drive business decisions. EDA is the first step in this process, where you will learn to ask the right questions and visualize data to uncover patterns, trends, and anomalies.

This Exploratory Data Analysis task in the context of retail at The Spark Foundation is an excellent opportunity for you to apply your data science and business analytics knowledge to a meaningful cause while developing valuable skills for your future career in the field